requirements.txt
pandas==2.2.3
numpy==2.2.6
matplotlib==3.10.3
opencv-python==4.11.0.86
Pillow==11.2.1
reportlab==4.4.1
streamlit==1.45.1
PyMuPDF==1.26.0
openpyxl==3.1.5
pytest==8.3.5

